seperate models into their own file
